pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "camoufox>=0.4.11",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
